+++ src/pages/AddSite.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import axios from 'axios';
-import "./../styles/AddSite.css";
+import "./styles/AddSite.css";
 import { useNavigate } from 'react-router-dom';
 
 // { websites, setWebsites }
