+++ src/pages/CreateCompany.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { useAuth0 } from "@auth0/auth0-react";
-import "./../styles/CreateCompany.css";
+import "./styles/CreateCompany.css";
 import { useNavigate } from "react-router-dom";
 
 function CreateCompany() {
