+++ src/pages/SiteDashboard.jsx
@@ -4,7 +4,7 @@ import MyCompany from "./site-pages/MyCompany";
 import Queries from "./site-pages/Queries";
 import IMPS from "./site-pages/Imps";
 import APIKey from "./site-pages/ApiKey";
-import "./../styles/SiteDashboard.css";
+import "./styles/SiteDashboard.css";
 import axios from "axios";
 import Loader from "./../components/Loader";
